WIP : 청크드를 처리하기 위한 노력
