+++ include/request_handler.hpp
@@ -5,7 +5,7 @@
 #include "datas.hpp"
 #include "webserv.hpp"
 
-t_error RequestHandler(size_t msg_len, void* udata, char* msg);
+int RequestHandler(struct kevent* curr_event);
 t_error convert_uri(std::string rq_uri,
                     const std::map<std::string, t_loc*>& location_config,
                     s_client_type& client);
