+++ storage/log/access.log
@@ -1,0 +1,19 @@
+127.0.0.1 [22/Mar/2023:22:51:31 + 0900] HTTP/1.1  GET 200 Ok [Path : ./storage/static/index.html]
+127.0.0.1 [01/Jan/1970:09:00:00 + 0900] HTTP/1.1  GET 200 Ok [Path : ./storage/static/asset/style.css]
+127.0.0.1 [22/Mar/2023:22:51:31 + 0900] HTTP/1.1  GET 200 Ok [Path : ./storage/static/asset/home.png]
+127.0.0.1 [22/Mar/2023:22:51:32 + 0900] HTTP/1.1  GET 200 Ok [Path : ./storage/static/asset/favicon_io/favicon-32x32.png]
+127.0.0.1 [22/Mar/2023:22:51:38 + 0900] HTTP/1.1  GET 200 Ok [Path : ./storage/static/index.html]
+127.0.0.1 [01/Jan/1970:09:00:00 + 0900] HTTP/1.1  GET 200 Ok [Path : ./storage/static/asset/style.css]
+127.0.0.1 [22/Mar/2023:22:51:38 + 0900] HTTP/1.1  GET 200 Ok [Path : ./storage/static/asset/wallpaper.jpeg]
+127.0.0.1 [01/Jan/1970:09:00:00 + 0900] HTTP/1.1  GET 200 Ok [Path : ./storage/static/asset/home.png]
+127.0.0.1 [22/Mar/2023:22:51:39 + 0900] HTTP/1.1  GET 200 Ok [Path : ./storage/static/asset/favicon_io/favicon-32x32.png]
+127.0.0.1 [22/Mar/2023:22:52:23 + 0900] HTTP/1.1  GET 200 Ok [Path : ./storage/static/test_post.html]
+127.0.0.1 [01/Jan/1970:09:00:00 + 0900] HTTP/1.1  GET 200 Ok [Path : ./storage/static/asset/favicon_io/favicon-32x32.png]
+127.0.0.1 [22/Mar/2023:22:54:52 + 0900] HTTP/1.1  GET 200 Ok [Path : ./storage/static/test_post.html]
+127.0.0.1 [22/Mar/2023:22:55:50 + 0900] HTTP/1.1  GET 200 Ok [Path : ./storage/static/asset/favicon_io/site.webmanifest]
+127.0.0.1 [22/Mar/2023:23:05:41 + 0900] HTTP/1.1  GET 200 Ok [Path : ./storage/static/index.html]
+127.0.0.1 [01/Jan/1970:09:00:00 + 0900] HTTP/1.1  GET 200 Ok [Path : ./storage/static/asset/style.css]
+127.0.0.1 [01/Jan/1970:09:00:00 + 0900] HTTP/1.1  GET 200 Ok [Path : ./storage/static/asset/home.png]
+127.0.0.1 [01/Jan/1970:09:00:00 + 0900] HTTP/1.1  GET 200 Ok [Path : ./storage/static/asset/favicon_io/favicon-32x32.png]
+127.0.0.1 [22/Mar/2023:23:05:45 + 0900] HTTP/1.1  GET 200 Ok [Path : ./storage/static/test_post.html]
+127.0.0.1 [01/Jan/1970:09:00:00 + 0900] HTTP/1.1  GET 200 Ok [Path : ./storage/static/asset/favicon_io/favicon-32x32.png]
