+++ storage/static/test_post.html
@@ -37,7 +37,7 @@
 	<form method="POST" action="/post" name="file_upload" enctype="multipart/form-data">
 		<fieldset style="width:400px;">
 		<label for="file" style="margin:10px 0px 10px 0px;">업로드할 사진을 선택하세요(png)</label>
-	  	<input type="file" id="file" style="margin:10px 0px 10px 0px;" accept=".png">
+	  	<input type="file" name="file" style="margin:10px 0px 10px 0px;" accept=".png">
 		<input type="submit" value="upload">
 		</fieldset>
 	</form>
